requirements-dev.txt
pydevd-pycharm
django-debug-toolbar
